+++ packages/web3-eth/src/utils/reject_if_block_timeout.ts
@@ -164,8 +164,11 @@ export async function rejectIfBlockTimeout(
 	const provider: Web3BaseProvider = web3Context.requestManager.provider as Web3BaseProvider;
 	let callingRes: [Promise<never>, ResourceCleaner];
 	const starterBlockNumber = await getBlockNumber(web3Context, NUMBER_DATA_FORMAT);
-	// TODO: once https://github.com/web3/web3.js/issues/5521 is implemented, remove checking for `enableExperimentalFeatures`
-	if (provider.supportsSubscriptions() && web3Context.enableExperimentalFeatures) {
+	// TODO: once https://github.com/web3/web3.js/issues/5521 is implemented, remove checking for `enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout`
+	if (
+		provider.supportsSubscriptions() &&
+		web3Context.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout
+	) {
 		callingRes = await resolveBySubscription(web3Context, starterBlockNumber, transactionHash);
 	} else {
 		callingRes = resolveByPolling(web3Context, starterBlockNumber, transactionHash);